examples/getting-started/pyproject.toml
[project]
name = "getting-started"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "bnbagent[server]",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "httpx>=0.25.0",
]

[tool.uv.sources]
bnbagent = { path = "../..", editable = true }
